Change type of cost related fields to Decimal

+++ travelsBackend/texpenses/models/models.py
@@ -20,7 +20,7 @@ from texpenses.validators import (
     start_end_date_validator)
 import googlemaps
 from geopy.geocoders import Nominatim
-
+from decimal import Decimal
 
 def update_instance(instance, updated_fields):
     for field, value in updated_fields.iteritems():
@@ -199,11 +199,11 @@ class Accommodation(md.Model):
     An abstract model that represents the accommodation related info
     """
 
-    accommodation_cost = md.FloatField(
+    accommodation_cost = md.DecimalField(max_digits=8, decimal_places=3,
         blank=False, default=0.0, validators=[MinValueValidator(0.0)])
     accommodation_default_currency = md.CharField(
         max_length=3, blank=False, default=settings.DEFAULT_CURRENCY)
-    accommodation_local_cost = md.FloatField(
+    accommodation_local_cost = md.DecimalField(max_digits=8, decimal_places=3,
         blank=False, default=0.0, validators=[MinValueValidator(0.0)])
     accommodation_local_currency = md.CharField(
         max_length=3, blank=True, choices=common.CURRENCIES)
@@ -222,7 +222,7 @@ class Transportation(md.Model):
     """
     An abstract model that represents the transportation related info
     """
-    transportation_cost = md.FloatField(
+    transportation_cost = md.DecimalField(max_digits=8, decimal_places=3,
         blank=False, default=0.0, validators=[MinValueValidator(0.0)])
     transportation_default_currency = md.CharField(
         max_length=3, blank=False, default=settings.DEFAULT_CURRENCY)
@@ -720,12 +720,12 @@ class ParticipationInfo(md.Model):
     """
     An abstract model that represents the participation cost related info
     """
-    participation_cost = md.FloatField(
+    participation_cost = md.DecimalField(max_digits=8, decimal_places=3,
         blank=False, default=0.0, validators=[MinValueValidator(0.0)])
 
     participation_default_currency = md.CharField(
         max_length=3, blank=False, default=settings.DEFAULT_CURRENCY)
-    participation_local_cost = md.FloatField(
+    participation_local_cost = md.DecimalField(max_digits=8, decimal_places=3,
         blank=True, default=0.0, validators=[MinValueValidator(0.0)])
     participation_local_currency = md.CharField(
         max_length=3, blank=True, choices=common.CURRENCIES)
@@ -744,14 +744,15 @@ class AdditionalCosts(md.Model):
     """
     An abstract model that represents the additional costs related info
     """
-    additional_expenses_initial = md.FloatField(
-        blank=False, default=0.0, validators=[MinValueValidator(0.0)])
+    additional_expenses_initial = md.DecimalField(
+        max_digits=8, decimal_places=3, blank=False, default=0.0,
+        validators=[MinValueValidator(0.0)])
     additional_expenses_default_currency = md.CharField(
         max_length=3, blank=False, default=settings.DEFAULT_CURRENCY)
     additional_expenses_initial_description = md.CharField(
         max_length=400, blank=True, null=True)
 
-    additional_expenses = md.FloatField(
+    additional_expenses = md.DecimalField(max_digits=8, decimal_places=3,
         blank=False, default=0.0, validators=[MinValueValidator(0.0)])
     additional_expenses_local_currency = md.CharField(
         max_length=3, blank=False, default=settings.DEFAULT_CURRENCY)
@@ -828,7 +829,7 @@ class Petition(SecretarialInfo, ParticipationInfo, AdditionalCosts):
     travel_files = md.FileField(upload_to=common.user_directory_path,
                                 null=True, blank=True)
 
-    total_cost_manual = md.FloatField(
+    total_cost_manual = md.DecimalField(max_digits=8, decimal_places=3,
         blank=False, default=0.0, validators=[MinValueValidator(0.0)])
     total_cost_change_reason = md.CharField(max_length=1000, blank=True,
                                             null=True)
@@ -1149,8 +1150,9 @@ class Petition(SecretarialInfo, ParticipationInfo, AdditionalCosts):
                 self.status <= self.APPROVED_BY_PRESIDENT) else (
                     self.additional_expenses)
 
-        return sum([compensation_cost_sum, additional_expenses,
+        return sum([Decimal(compensation_cost_sum), additional_expenses,
                     self.transportation_cost_to_be_compensated()])
+
 
 
     def total_cost_calculated(self):
